==> explore_figure/Fig5_Macrophage_journal_replication/03_Make_new_columns_and_save_RDS.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(tidyverse)
library(here)
library(SingleCellExperiment)
```

```{r}
macrophage <- readRDS(here('explore/RDS/macrophage_journal_replicaton.RDS'))
```

```{r}
DimPlot(macrophage, group.by = "Diagnosis", label = T)
DimPlot(macrophage, group.by = "seurat_clusters", label = T)
```

```{r}
cluster23 <- subset(macrophage, subset = seurat_clusters == 23)
```

```{r}
DimPlot(cluster23, group.by = "seurat_clusters", label = T)
```



```{r}
macrophage@meta.data$new_submain <- ifelse(macrophage@meta.data$seurat_clusters %in% c(16, 21, 8, 6, 10, 22, 15, 24), "FCN1+ Monocyte", "NA")
macrophage@meta.data$new_submain <- ifelse(macrophage@meta.data$seurat_clusters %in% c(0, 25, 4, 19, 7, 14), "SPP1+ Macrophage", macrophage@meta.data$new_submain)
macrophage@meta.data$new_submain <- ifelse(macrophage@meta.data$seurat_clusters %in% c(12, 11, 2, 1, 17, 13, 3, 20), "FABP4+ Macrophage", macrophage@meta.data$new_submain)
macrophage@meta.data$new_submain <- ifelse(macrophage@meta.data$seurat_clusters %in% c(5, 23), "Mixed Macrophage", macrophage@meta.data$new_submain)
macrophage@meta.data$new_submain <- ifelse(macrophage@meta.data$seurat_clusters %in% c(18), "SFTPC+ Macrophage", macrophage@meta.data$new_submain)
```

# cluster 9 : ipf only 
# remove cluster 22, 9 

```{r}
macrophage <- subset(macrophage, seurat_clusters != 22)
macrophage <- subset(macrophage, seurat_clusters != 9)
macrophage <- ScaleData(macrophage)
macrophage <- RunPCA(macrophage)
macrophage <- RunUMAP(macrophage, dims = 1:30)
```

```{r}
DimPlot(macrophage, group.by = "new_submain", label = T)
DimPlot(macrophage, group.by = "seurat_clusters", label = T)
```

```{r}
saveRDS(macrophage, file = "macrophage_journal_replicaton2.RDS")
```
