==> explore_figure/Fig5_Macrophage_journal_replication/05_Violin_new_columns_diagnosis_tag.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Seurat)
library(tidyverse)
library(here)
library(SingleCellExperiment)
```

```{r}
macrophage <- readRDS(here('explore/RDS/macrophage_journal_replicaton2.RDS'))
```

```{r}
macrophage$row <- substr(rownames(macrophage@meta.data), 1, 6)
head(macrophage)
```

```{r}
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01157"), "HC1", "NA")
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01174"), "HC2", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01365"), "HC3", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01366"), "HC4", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01367"), "HC5", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("HD65_A", "HD65_C", "HD65_G", "HD65_T"), "HC6", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("HD66_A", "HD66_C", "HD66_G", "HD66_T"), "HC7", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("HD67_A", "HD67_C", "HD67_G", "HD67_T"), "HC8", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("HD68_A", "HD68_C", "HD68_G", "HD68_T"), "HC9", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F00409"), "HC10", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("HD70_A", "HD70_C", "HD70_G", "HD70_T"), "HC11", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01394"), "HC12", macrophage@meta.data$Diagnosis_tag)

macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F00431"), "I1", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01302"), "I2", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01303"), "I3", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01214"), "I4", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01172"), "I5", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01173"), "I6", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01379"), "I7", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01380"), "I8", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("ILD53_"), "I9", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("ILD59-"), "I10", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("ILD60-"), "I10", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("ILD61-"), "I11", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("ILD63_"), "I12", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01390"), "I13", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01391"), "I14", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("F01392"), "I15", macrophage@meta.data$Diagnosis_tag)

macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("obj1_A", "obj1_C", "obj1_G", "obj1_T"), "C1", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("obj2_A", "obj2_C", "obj2_G", "obj2_T"), "C2", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("obj3_A", "obj3_C", "obj3_G", "obj3_T"), "C3", macrophage@meta.data$Diagnosis_tag)
macrophage@meta.data$Diagnosis_tag <- ifelse(macrophage@meta.data$row %in% c("obj4_A", "obj4_C", "obj4_G", "obj4_T"), "C4", macrophage@meta.data$Diagnosis_tag)


head(macrophage)
unique(macrophage$Diagnosis_tag)
```

```{r}
saveRDS(macrophage, file = "macrophage_journal_replicaton3.RDS")
```
