==> explore_figure/Fig5_Macrophage_journal_replication/10_Monocle_splitby_diagnosis.Rmd <==
---
title: "Untitled"
output: html_document
---

```{r setup, include=FALSE}
library(Signac)
library(Seurat)
library(SeuratWrappers)
library(monocle3)
library(Matrix)
library(ggplot2)
library(patchwork)
library(here)
```

```{r}
macrophage <- readRDS(here('explore/RDS/macrophage_new_marker.RDS'))
```

```{r}
control <- subset(macrophage, subset = Diagnosis == "Control")
covid <- subset(macrophage, subset = Diagnosis == "COVID")
ipf <- subset(macrophage, subset = Diagnosis == "IPF")
```

# Building tranjectories with monocle3 
```{r}
control.cds <- as.cell_data_set(control)
control.cds <- cluster_cells(cds = control.cds, reduction_method = "UMAP")
control.cds <- learn_graph(control.cds, use_partition = TRUE)

covid.cds <- as.cell_data_set(covid)
covid.cds <- cluster_cells(cds = covid.cds, reduction_method = "UMAP")
covid.cds <- learn_graph(covid.cds, use_partition = TRUE)

ipf.cds <- as.cell_data_set(ipf)
ipf.cds <- cluster_cells(cds = ipf.cds, reduction_method = "UMAP")
ipf.cds <- learn_graph(ipf.cds, use_partition = TRUE)
```

# root cells selection
```{r}
root_subset_control <- subset(control, subset = seurat_clusters == 10)
root_subset_control <- root_subset_control@meta.data$X
head(root_subset_control)

root_subset_covid <- subset(covid, subset = seurat_clusters == 10)
root_subset_covid <- root_subset_covid@meta.data$X
head(root_subset_covid)

root_subset_ipf <- subset(ipf, subset = seurat_clusters == 10)
root_subset_ipf <- root_subset_ipf@meta.data$X
head(root_subset_ipf)
```

# order cells 
```{r}
control.cds <- order_cells(control.cds, reduction_method = "UMAP", root_cells = root_subset_control)
covid.cds <- order_cells(covid.cds, reduction_method = "UMAP", root_cells = root_subset_covid)
ipf.cds <- order_cells(ipf.cds, reduction_method = "UMAP", root_cells = root_subset_ipf)
```

# plot trajecotires colored by pesudotime 
```{r}
plot_cells(
  cds = control.cds, 
  color_cells_by = "pseudotime",
  show_trajectory_graph = TRUE
)

plot_cells(
  cds = covid.cds, 
  color_cells_by = "pseudotime",
  show_trajectory_graph = TRUE
)

plot_cells(
  cds = ipf.cds, 
  color_cells_by = "pseudotime",
  show_trajectory_graph = TRUE
)
```

```{r}
control.cds <- AddMetaData(
  object = control,
  metadata = control.cds@principal_graph_aux@listData$UMAP$pseudotime,
  col.name = "Control"
)


FeaturePlot(control.cds, c("Control"), pt.size = 0.1) & scale_color_viridis_c()
#ggsave(here('explore_figure/Fig3/Control_monocle.pdf'))
```

```{r}
covid.cds <- AddMetaData(
  object = covid,
  metadata = covid.cds@principal_graph_aux@listData$UMAP$pseudotime,
  col.name = "COVID"
)


FeaturePlot(covid.cds, c("COVID"), pt.size = 0.1) & scale_color_viridis_c()
#ggsave(here('explore_figure/Fig3/COVID_monocle.pdf'))
```

```{r}
ipf.cds <- AddMetaData(
  object = ipf,
  metadata = ipf.cds@principal_graph_aux@listData$UMAP$pseudotime,
  col.name = "IPF"
)


FeaturePlot(ipf.cds, c("IPF"), pt.size = 0.1) & scale_color_viridis_c()
#ggsave(here('explore_figure/Fig3/IPF_monocle.pdf'))
```
